Check for broken build due to invalid devices.png (non-LFS clone)
Cloning the repo without LFS, leaves a text file with a hash in place of
icons/devices.png instead of the real file. This causes a broken build
with spnavcfg crashing on startup.

This was documented from the start in the README.md file, but since
apparently nobody bothers to read that, I added an explicit broken build
detection and bailout at startup, with an informative error message, as
was suggested in pull request #26 discussion.